api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
google-cloud-pubsub==2.22.0
google-cloud-storage==2.18.2
